+++ note_maker/services/UserServices.py
@@ -1,11 +1,12 @@
 from flask import request
 from flask_restful import Resource
 
-from note_maker import session
+from note_maker import session, auth, app
 from note_maker.models import User, Note
 from note_maker.schemas import (
     user_schema, user_list_schema, note_list_schema)
 from note_maker.services.Exceptions import Message
+from werkzeug.security import check_password_hash
 
 
 class UserService(Resource):
@@ -31,9 +32,13 @@ class UserService(Resource):
 
         return Message.successful('created', 201)
 
+    @auth.login_required
     def put(self, user_id):
         request_data = request.get_json()
         user = session.query(User).get(user_id)
+
+        if user.email != auth.current_user().email:
+            return Message.auth_failed()
 
         if user is None:
             return Message.instance_not_exist()
@@ -54,8 +59,13 @@ class UserService(Resource):
             return Message.instance_not_exist()
         return user_schema.dump(user), 200
 
+    @auth.login_required
     def delete(self, user_id):
         user = session.query(User).get(user_id)
+
+        if user.email != auth.current_user().email:
+            return Message.auth_failed()
+
         if user is None:
             return Message.instance_not_exist()
         session.delete(user)
@@ -81,3 +91,16 @@ class UserNotesService(Resource):
 
         user_note_list = user.notes
         return note_list_schema.dump(user_note_list)
+
+
+@auth.verify_password
+def verify_password(login, password):
+    user = session.query(User).filter(User.email == login).first()
+    if user and check_password_hash(user.password, password):
+        return user
+
+
+@app.route('/index', methods=['GET'])
+@auth.login_required
+def index():
+    return "Hello, {}!".format(auth.current_user())
